Add files via upload
Lua code that is a path finding algorithm to direct a "turtle" from the computer craft minecraft mod to go to a locaiton and carve of a quarry size that the user specifies. Currently working on creating server side code that would allow an army of them to excavate the whole world.